+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+* `flexberry-groupedit` component:
+    * Add check all at page button.
+
 ### Changed
 * `flexberry-lookup` component:
     * Dropdown mode uses projection from `projection` property.